Move ProjectionTy methods to extension trait
